Code cleaning

+++ commands/uim/commands/classes/caches/clear.d
@@ -40,7 +40,7 @@ class DCacheClearCommand : DCommand {
   }
 
   // Implement this method with your command`s logic.
-  override ulong execute(Json[string] arguments, IConsole aConsole = null) {
+  override size_t execute(Json[string] arguments, IConsole aConsole = null) {
     string engineName = to!string(commandArguments.getArgument("engine"));
     try {
       aConsoleIo.writeln("Clearing {engineName}");
