+++ commands/uim/commands/classes/caches/cleargroup.d
@@ -41,7 +41,7 @@ class DCacheClearGroupCommand : DCommand {
   }
 
   // Clears the cache group
-  override ulong execute(Json[string] arguments, IConsole aConsole = null) {
+  override size_t execute(Json[string] arguments, IConsole aConsole = null) {
     auto anGroup = to!string(commandArguments.getArgument("group"));
     try {
       anGroupConfigs = Cache.groupConfigs(anGroup);
